feat: replace modal
