--- src/components/Modal/Modal.module.scss
@@ -1,35 +1,0 @@
-@use '@/styles/variables.scss' as vars;
-@use '@/styles/mixins.scss' as mixins;
-
-.modaloverlay {
-  position: fixed;
-  z-index: 1000;
-  top: 0;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  background-color: vars.$modal;
-}
-
-.modalcontent {
-  position: relative;
-  overflow: auto;
-  max-width: 90%;
-  max-height: 90%;
-  padding: 1rem;
-  border-radius: 8px;
-}
-
-.modalclose {
-  cursor: pointer;
-  position: absolute;
-  z-index: 1010;
-  top: 0.5rem;
-  right: 1.2rem;
-  border: none;
-  font-size: 3rem;
-  background: none;
-}
