+++ src/shared/ui/Modal/Modal.module.scss
@@ -1,21 +1,35 @@
-@use '@/styles/variables.scss' as vars;
+@use '@/styles/variables' as vars;
+@use '@/styles/mixins' as mixins;
 
-.backdrop {
+.modaloverlay {
   position: fixed;
   z-index: 1000;
-  inset: 0;
+  top: 0;
+  left: 0;
   display: flex;
   align-items: center;
   justify-content: center;
-  background: rgb(0 0 0 / 40%);
+  width: 100%;
+  height: 100%;
+  background-color: vars.$modal;
 }
 
-.modal {
+.modalcontent {
   position: relative;
-  width: 100%;
-  max-width: 500px;
-  padding: 2rem;
-  border-radius: 12px;
-  background: vars.$color-primary100;
-  box-shadow: 0 20px 40px rgb(0 0 0 / 20%);
+  overflow: auto;
+  max-width: 90%;
+  max-height: 90%;
+  padding: 1rem;
+  border-radius: 8px;
+}
+
+.modalclose {
+  cursor: pointer;
+  position: absolute;
+  z-index: 1010;
+  top: 0.5rem;
+  right: 1.2rem;
+  border: none;
+  font-size: 3rem;
+  background: none;
 }
